+++ wantPlant/app/src/main/java/com/example/wantplant/ui/main/login/LoginActivity.kt
@@ -1,0 +1,27 @@
+package com.example.wantplant.ui.main.login
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.example.wantplant.R
+import com.example.wantplant.databinding.ActivityLoginBinding
+import com.example.wantplant.databinding.ActivityMainBinding
+import com.example.wantplant.databinding.FragmentLandingBinding
+import com.example.wantplant.ui.main.book.BookFragment
+import com.example.wantplant.ui.main.book.LandingFragment
+import com.example.wantplant.ui.main.book.LandingPageFragment
+import com.example.wantplant.ui.main.garden.GardenFragment
+import com.example.wantplant.ui.main.profile.ProfileFragment
+import com.example.wantplant.ui.main.water.month.WaterMonthFragment
+
+class LoginActivity : AppCompatActivity() {
+
+    private lateinit var binding : ActivityLoginBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        binding = ActivityLoginBinding.inflate(layoutInflater)
+
+        setContentView(binding.root)
+    }
+}
